% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{analyze_dose_response}
\alias{analyze_dose_response}
\title{Process dose response curves}
\usage{
analyze_dose_response(
  plate_data,
  plot_dose_response,
  directory_curves,
  cores_n
)
}
\arguments{
\item{plate_data}{Processed plate data}

\item{plot_dose_response}{Logical. If TRUE, generates dose-response curve plots}

\item{directory_curves}{String. Directory path for saving dose-response curve plots}

\item{cores_n}{Integer. Number of cores for parallel processing}
}
\value{
Data frame with curve fitting results
}
\description{
Process dose response curves
}
