% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{calc_pop_sd}
\alias{calc_pop_sd}
\title{Calculate population standard deviation}
\usage{
calc_pop_sd(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Population standard deviation
}
\description{
Calculate population standard deviation
}
