% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{calc_robust_zfactor}
\alias{calc_robust_zfactor}
\title{Calculate robust Z-factor}
\usage{
calc_robust_zfactor(x, y)
}
\arguments{
\item{x, y}{Numeric vectors to compare}
}
\value{
Robust Z-factor value
}
\description{
Calculate robust Z-factor
}
