% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{calc_ssmd}
\alias{calc_ssmd}
\title{Calculate SSMD (Strictly Standardized Mean Difference)}
\usage{
calc_ssmd(x, y)
}
\arguments{
\item{x, y}{Numeric vectors to compare}
}
\value{
SSMD value
}
\description{
Calculate SSMD (Strictly Standardized Mean Difference)
}
