% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{calculate_all}
\alias{calculate_all}
\title{Fit dose-response curves and get residuals}
\usage{
calculate_all(
  dose,
  inhibition_percent,
  dataset,
  drug_name,
  WELL = WELL,
  plot_dose_response = plot_dose_response,
  directory_curves = directory_curves
)
}
\arguments{
\item{dose}{Numeric vector of drug concentrations}

\item{inhibition_percent}{Numeric vector of inhibition percentages}

\item{dataset}{Character vector identifying different datasets}

\item{drug_name}{Character string of drug name}

\item{WELL}{Character vector of well identifiers}

\item{plot_dose_response}{Logical; if TRUE, generates dose-response curve plots}

\item{directory_curves}{String; directory path for saving dose-response curve plots}
}
\value{
Data frame with fitted values and analysis metrics
}
\description{
Fit dose-response curves and get residuals
}
