% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{create_combined_plot}
\alias{create_combined_plot}
\title{Create combined visualization}
\usage{
create_combined_plot(plate_data, plate_data_raw, file_path)
}
\arguments{
\item{plate_data}{Processed plate data with residuals}

\item{plate_data_raw}{Raw processed plate data with controls preserved}

\item{file_path}{Output file path}
}
\value{
None, saves plot to file
}
\description{
Create combined visualization
}
