% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{create_plate_heatmap}
\alias{create_plate_heatmap}
\title{Generate plate heatmap}
\usage{
create_plate_heatmap(plate_data, data_type)
}
\arguments{
\item{plate_data}{Processed plate data}

\item{data_type}{String; type of data to visualize ("inhibition_percent" or "residuals_norm")}
}
\value{
ggplot object
}
\description{
Generate plate heatmap
}
