% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{create_plate_labels}
\alias{create_plate_labels}
\title{Create plate row labels}
\usage{
create_plate_labels(n)
}
\arguments{
\item{n}{Number of rows}
}
\value{
Character vector of row labels
}
\description{
Create plate row labels
}
