% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{create_row_scatter}
\alias{create_row_scatter}
\title{Create scatter plot of inhibition by row}
\usage{
create_row_scatter(plate_data)
}
\arguments{
\item{plate_data}{Processed plate data}
}
\value{
ggplot object
}
\description{
Create scatter plot of inhibition by row
}
