% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{process_plate}
\alias{process_plate}
\title{Process plate data}
\usage{
process_plate(plate_data)
}
\arguments{
\item{plate_data}{Data frame containing plate data}
}
\value{
List containing processed data and quality metrics
}
\description{
Process plate data
}
