% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateQC.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers using IQR method}
\usage{
remove_outliers(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector with outliers replaced by NA
}
\description{
Removes outliers from a numeric vector using the Interquartile Range method
}
